Add files via upload
This is my first project in Deep Learning. I used a simple and straight-forward approach to classifying the images as parasitic or normal. I achieved an accuracy of just over 80% which is not very good. I would love any comments and/or help to really understand how to optimize my algorithm's accuracy to the high 90s.